Convert cache source (#4)
* # gps.py
* rearranged the code into more appropriate sections in order to make
specific functions easier to find.
* converted caching mechanism to sql lite
* added a generic class for api requests. the source can be either
google or yelp (yelp not yet fully implemented) but is structured in a
way that allows adding future source locations more easily. the calling
and parsing functions are assigned to the api request class object upon
or after instantiation
* added enums for the user to select which source to use

* # gps.py
* removed IGNORED_PLACE_TYPES and other Google Maps globals. The mapping
became more complicated between Goggle and Yelp so I created a new
notebook - place_categories.ipynb - that creates and uploades the
mapping info to Synapse

* converted the caching mechanism from a csv to sqlite to increase
efficiency.

* converted the api requesting mechanism to a more functional approach
so that we can add in other apis more easily. i've emailed both citygrid
and foursquare to see what they can do for us in regards to the API.
neither of their licenses allow for research purposes on the basic
licence and require enterprise.

* added a parameter in the request_nearby_places signature to skip the
cache and pull the most recent content from the designated api source.
Doing this required an additional function - put_to_cache - that checks
the cache for existing content with the same key and removes those rows
before inserting the new content.

* added a feature to the cluster metrics that calculates the mean time
between each cluster visit

* no longer checks or stores API keys. must be supplied with each
request. see the demo for a how to

# weather.py
* no longer checks or stores API keys. must be supplied with each
request. see the demo for a how to

* removed the download of existing weather cache as it was overwriting
any cached items that may have been queried since the cache upload to
synapse.

* added geo_distance to this script as well. would have rather included
the gps.py in the module init but doing this removes external
dependencies as required