# __init__.py included utils to deduplicate code
# gps.py
* removed imports that are imported by including all of utils
* moved duplicated code between gps.py and weather.py into utils.py
* moved zips, yelp_mappings, and gmap_mappings into github directory as
required data files so we don't need to login into Synapse. Synapse
requires login for download. Public access can only view the data
* added lines in both extract_cluster_centers and gps_dbscan that use
newly added methods to check and convert from dataframe if required

# utils.py
* moved the duplicated code from gps.py and weather.py in here for more
concise testing.

# weather.py
* removed imports and other methods brought in by utils

# moved requirements.txt into notebooks directory to make sure users
understand that this file is a list of dependencies only for my work
that generated content for the library and not the library itself

# test_gps.py
* added tests for new methods described above