# weather.ipynb  * updated to match modified weather scripts
 # weather.py
* separated date and time into two different columns for sqlite backend
* changed session_scope to include an argument so that we can connect
to a test_cache for unit test
* changed weather_request to allow any assortment of lat, lon or zipcode
pairs with date. it is no longer reliant on the named tuple defined in
the module but extra methods were added to parse the different types and
possible orderings. a kwargs argument was also added that is only used
for testing purposes
* removed the weekly summary option until we can define a better scope
for summarization
* other minor adjustments made for minimizing unit testing

# test_utils.py and test_weather.py
* added small unit tests to help me make the changes defined above